Rename optee_msg_fs.h to optee_msg_supplicant.h
